構造体の別ファイル作成
