+++ WhatdoApp/ContentView.swift
@@ -7,17 +7,6 @@
 
 import SwiftUI
 
-struct Task: Identifiable {
-    let id = UUID()
-    var title: String
-    var checked: Bool
-    
-    init(title: String, checked: Bool) {
-        self.title = title
-        self.checked = checked
-    }
-}
-
 struct ContentView: View {
     var userData = UserData()
